Refactor packages for internal view annotation files
